Refactor: 더보기 버튼 상단으로 이동, 좋아요 버튼 고정 #132
