+++ src/components/PlaylistDetail/MusicPlayBar.jsx
@@ -1,30 +1,20 @@
-import React, { useState, useEffect } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
-import { useRecoilState, useRecoilValue } from 'recoil';
+import React, { useState } from 'react';
+
 import styled from 'styled-components';
 
-import { userInfoAtom, PlayListAtom } from '../../library/atom';
-import {
-  useDeletePlaylist,
-  useLikePlaylist,
-} from '../../hooks/queries/usePlaylist';
+import { useLikePlaylist } from '../../hooks/queries/usePlaylist';
 
 import MiniModal from '../common/Modal/MiniModal';
 
 import PlayIcon from '../../img/play-icon.svg';
 import PauseIcon from '../../img/pause-Icon.svg';
-import MoreIcon from '../../img/more-icon.svg';
 import ShareIcon from '../../img/share-icon.svg';
 import LikeIcon from '../../img/like-icon.svg';
 import LikeActiveIcon from '../../img/like-active-icon.svg';
 
 export default function MusicPlayBar(props) {
-  const navigate = useNavigate();
-  const myId = useRecoilValue(userInfoAtom).id;
-  const { mutate: deletePlaylist } = useDeletePlaylist();
   const {
     playlist,
-    userId,
     playlistId,
     playing,
     setPlaying,
@@ -32,11 +22,7 @@ export default function MusicPlayBar(props) {
     setPause,
     setCurrMusic,
   } = props;
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const toggleModal = () => {
-    if (isModalOpen === false) setIsModalOpen(true);
-    else setIsModalOpen(false);
-  };
+
   const { mutate: likePlaylist } = useLikePlaylist();
 
   const handlePlayBtn = () => {
@@ -47,16 +33,6 @@ export default function MusicPlayBar(props) {
     } else {
       setPause(!pause);
     }
-  };
-
-  const handleDeleteBtnClick = () => {
-    const id = playlistId;
-    deletePlaylist(id, {
-      onSuccess: () => {
-        alert('플레이리스트가 정상적으로 삭제되었습니다.');
-        navigate(-1);
-      },
-    });
   };
 
   const handleShareBtnClick = () => {
@@ -80,32 +56,13 @@ export default function MusicPlayBar(props) {
           alt='재생/멈춤 버튼'
         />
       </PlayBtn>
-      <MoreBtnBox>
-        {userId === myId ? (
-          <button onClick={toggleModal}>
-            <img src={MoreIcon} alt='더보기 버튼' />
-          </button>
-        ) : (
-          <LikeBtn onClick={handleLikeBtnClick}>
-            <img
-              src={playlist.like_playlist ? LikeActiveIcon : LikeIcon}
-              alt='좋아요'
-            />
-            <p>{playlist.like_count}</p>
-          </LikeBtn>
-        )}
-        {userId === myId && isModalOpen && (
-          <MiniModal>
-            <button onClick={handleDeleteBtnClick}>플리 삭제</button>
-            <Link
-              to={`/playlist/detail/${playlistId}/edit`}
-              state={{ id: playlistId }}
-            >
-              플리 수정
-            </Link>
-          </MiniModal>
-        )}
-      </MoreBtnBox>
+      <LikeBtn onClick={handleLikeBtnClick}>
+        <img
+          src={playlist.like_playlist ? LikeActiveIcon : LikeIcon}
+          alt='좋아요'
+        />
+        <p>{playlist.like_count}</p>
+      </LikeBtn>
     </PlayBarWrap>
   );
 }
@@ -117,7 +74,7 @@ const PlayBarWrap = styled.div`
   align-items: center;
   justify-content: center;
   gap: 35px;
-
+  box-shadow: 0px 2px 4px 0px rgba(0, 0, 0, 0.05);
   button img {
     vertical-align: middle;
   }
