+++ src/pages/PlaylistDetail/PlaylistDetail.jsx
@@ -39,6 +39,7 @@ export default function PlaylistDetail() {
   const { playlist, comments, music, user } = data;
   const musicList = music.map((obj) => obj.information);
   const musicLength = music.length;
+
   return (
     <>
       <PlaylistDetailWrap>
@@ -54,7 +55,6 @@ export default function PlaylistDetail() {
         )}
         <MusicPlayBar
           playlist={playlist}
-          userId={user.id}
           playlistId={playlistId}
           pause={pause}
           setPause={setPause}
